+++ app/src/main/java/com/nbc/messenger/ui/my/MyPageFragment.kt
@@ -1,10 +1,11 @@
-package com.nbc.messenger
+package com.nbc.messenger.ui.my
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
+import com.nbc.messenger.R
 import com.nbc.messenger.data.DataSource
 import com.nbc.messenger.databinding.FragmentMyPageBinding
 import com.nbc.messenger.model.ProfileImage
