+++ app/src/main/res/layout/fragment_contact_list.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ContactListFragment">
+    tools:context=".ui.main.ContactListFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/constView"
