+++ app/src/main/res/layout/fragment_datail.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".DetailFragment">
+    tools:context=".ui.detail.DetailFragment">
 
     <Button
         android:id="@+id/btn_detail_back"
