+++ app/src/main/res/layout/fragment_view_pager.xml
@@ -5,7 +5,7 @@
     android:id="@+id/frame"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ViewPagerFragment">
+    tools:context=".ui.main.ViewPagerFragment">
 
     <FrameLayout
         android:id="@+id/frameLayout2"
