+++ app/src/main/res/layout/fragment_view_pager_adapter.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ViewPagerAdapter">
+    tools:context=".ui.main.ViewPagerAdapter">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
